+++ src/main/java/com/elmakers/mine/bukkit/magic/Mage.java
@@ -1759,6 +1759,11 @@ public class Mage implements CostReducer, com.elmakers.mine.bukkit.api.magic.Mag
     public Entity getEntity() {
         return _entity.get();
     }
+    
+    @Override
+    public EntityData getEntityData() {
+        return entityData;
+    }
 
     @Override
     public LivingEntity getLivingEntity() {